Rearranged aux functions in more appropriate locations
